Seperate classes from main, fix place
I need you to test and optimize the serialization functions related to "places" so that I can finish the traveler class.
