+++ timePeriod.cpp
@@ -1,0 +1,3 @@
+#include "timePeriod.h"
+bool timePeriod::includesDate(Time t) {}
+bool timePeriod::includesPeriod() {}
